# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MeResponseDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"is_phone_verified": {
						"type": "boolean"
					},
					"is_email_verified": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"business": {
						"properties": {
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"is_phone_verified",
					"is_email_verified",
					"created_at",
					"roles",
					"business"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"is_phone_verified": {
						"type": "boolean"
					},
					"is_email_verified": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"is_phone_verified",
					"is_email_verified",
					"created_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMeResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/UserDto"
					},
					"email_verification_sent": {
						"type": "boolean"
					},
					"phone_verification_sent": {
						"type": "boolean"
					}
				},
				"required": [
					"message",
					"user",
					"email_verification_sent",
					"phone_verification_sent"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMeDto": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 100
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"phone": {
						"type": "string",
						"pattern": "^\\+?[1-9]\\d{1,14}$"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMyPasswordDto": {
				"properties": {
					"current_password": {
						"type": "string",
						"minLength": 1
					},
					"new_password": {
						"type": "string",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"
					}
				},
				"required": [
					"current_password",
					"new_password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InviteResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"invitation_id": {
						"type": "number",
						"format": "double"
					},
					"expires_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"message",
					"invitation_id",
					"expires_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InviteDto": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"role_id": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					}
				},
				"required": [
					"email",
					"role_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AcceptInviteResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"user": {
						"properties": {
							"email": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"email",
							"name",
							"id"
						],
						"type": "object"
					},
					"tokens": {
						"properties": {
							"refresh_token": {
								"type": "string"
							},
							"access_token": {
								"type": "string"
							}
						},
						"required": [
							"refresh_token",
							"access_token"
						],
						"type": "object"
					}
				},
				"required": [
					"message",
					"user",
					"tokens"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AcceptInviteDto": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					},
					"business_id": {
						"type": "integer",
						"format": "int32",
						"minimum": 1
					},
					"token": {
						"type": "string",
						"minLength": 1
					},
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 100
					},
					"phone": {
						"type": "string",
						"pattern": "^\\+?[1-9]\\d{1,14}$"
					},
					"password": {
						"type": "string",
						"minLength": 8,
						"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).*$"
					}
				},
				"required": [
					"email",
					"business_id",
					"token",
					"name",
					"phone",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserWithRolesDto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"is_phone_verified": {
						"type": "boolean"
					},
					"is_email_verified": {
						"type": "boolean"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"roles": {
						"items": {
							"properties": {
								"name": {
									"type": "string"
								},
								"id": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"name",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"is_phone_verified",
					"is_email_verified",
					"created_at",
					"roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetUsersResponseDto": {
				"properties": {
					"users": {
						"items": {
							"$ref": "#/components/schemas/UserWithRolesDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"users",
					"total",
					"page",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminUpdateUserResponseDto": {
				"properties": {
					"message": {
						"type": "string"
					},
					"user": {
						"properties": {
							"roles": {
								"items": {
									"properties": {
										"name": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"name",
										"id"
									],
									"type": "object"
								},
								"type": "array"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"roles",
							"name",
							"id"
						],
						"type": "object"
					}
				},
				"required": [
					"message",
					"user"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminUpdateUserDto": {
				"properties": {
					"name": {
						"type": "string",
						"minLength": 2,
						"maxLength": 100
					},
					"role_ids": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "einvoicebackend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateMeResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMeDto"
							}
						}
					}
				}
			}
		},
		"/users/me/password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMyPasswordDto"
							}
						}
					}
				}
			}
		},
		"/users/invite": {
			"post": {
				"operationId": "InviteUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InviteResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"OWNER",
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InviteDto"
							}
						}
					}
				}
			}
		},
		"/users/accept-invite": {
			"post": {
				"operationId": "AcceptInvite",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AcceptInviteResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptInviteDto"
							}
						}
					}
				}
			}
		},
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUsersResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"OWNER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserWithRolesDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"OWNER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "AdminUpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminUpdateUserResponseDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"OWNER",
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminUpdateUserDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}